Delete extra aggregate package
