+++ api-java/src/test/java/javaclasses/mealorder/c/aggregate/VendorCommandTest.java
@@ -18,7 +18,7 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
 
-package javaclasses.mealorder.c.definition;
+package javaclasses.mealorder.c.aggregate;
 
 import com.google.protobuf.Message;
 import io.spine.client.TestActorRequestFactory;
